refactor(packages/feature-flags): load flags ahead of time
All the feature flags will now be loaded ahead of time, then when
queried the client will just return the pre-loaded flag values. This is
more efficient given there aren't likely to be many flags, and this will
also allow us to reference the flags synchronously.
