+++ packages/feature-flags/index.js
@@ -1,2 +1,1 @@
-export * from './src/is-feature-active.js';
 export * from './src/feature-flag-client.js';
